Sketch-GUI-version1/requirements.txt
certifi==2021.5.30
charset-normalizer==2.0.4
colorutils==0.3.0
cycler==0.10.0
editdistance==0.5.3
future==0.18.2
idna==3.2
joblib==1.0.1
kiwisolver==1.3.2
matplotlib==3.4.3
mlxtend==0.19.0
numexpr==2.7.3
numpy==1.21.2
pandas==1.3.3
pandastable==0.12.2.post1
Pillow==8.3.2
plotly==5.3.1
psutil==5.8.0
psycopg2-binary==2.9.1
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2021.1
requests==2.26.0
scikit-learn==0.24.2
scipy==1.7.1
seaborn==0.11.2
six==1.16.0
sklearn==0.0
tenacity==8.0.1
threadpoolctl==2.2.0
tkinterhtml==0.7
urllib3==1.26.6
xlrd==2.0.1
